+++ it/server/src/test/java/com/walmartlabs/concord/it/server/UserManagementIT.java
@@ -33,7 +33,6 @@ public class UserManagementIT extends AbstractServerIT {
 
         // ---
 
-        setApiKey(cakr.getKey());
         userResource.delete(cur.getId());
     }
 }