+++ server/impl/src/main/java/com/walmartlabs/concord/server/user/RoleResourceImpl.java
@@ -1,21 +1,19 @@
 package com.walmartlabs.concord.server.user;
 
 import com.walmartlabs.concord.server.api.PerformedActionType;
-import com.walmartlabs.concord.server.api.security.Permissions;
 import com.walmartlabs.concord.server.api.user.CreateRoleResponse;
 import com.walmartlabs.concord.server.api.user.DeleteRoleResponse;
 import com.walmartlabs.concord.server.api.user.RoleEntry;
 import com.walmartlabs.concord.server.api.user.RoleResource;
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.authz.UnauthorizedException;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.apache.shiro.subject.Subject;
+import com.walmartlabs.concord.server.security.UserPrincipal;
 import org.sonatype.siesta.Resource;
 import org.sonatype.siesta.Validate;
 import org.sonatype.siesta.ValidationErrorsException;
 
 import javax.inject.Inject;
 import javax.inject.Named;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Named
@@ -31,18 +29,14 @@ public class RoleResourceImpl implements RoleResource, Resource {
     @Override
     @Validate
     public CreateRoleResponse createOrUpdate(RoleEntry entry) {
+        assertAdmin();
+
         String roleName = entry.getName();
 
         if (roleDao.exists(roleName)) {
-            assertPermissions(Permissions.ROLE_UPDATE_ANY,
-                    "The current user does not have permissions to update the specified role");
-
             roleDao.update(roleName, entry.getDescription(), entry.getPermissions());
             return new CreateRoleResponse(PerformedActionType.UPDATED);
         } else {
-            assertPermissions(Permissions.ROLE_CREATE_NEW,
-                    "The current user does not have permissions to create a new role");
-
             roleDao.insert(roleName, entry.getDescription(), entry.getPermissions());
             return new CreateRoleResponse(PerformedActionType.CREATED);
         }
@@ -55,8 +49,9 @@ public class RoleResourceImpl implements RoleResource, Resource {
 
     @Override
     @Validate
-    @RequiresPermissions(Permissions.ROLE_DELETE_ANY)
     public DeleteRoleResponse delete(String name) {
+        assertAdmin();
+
         if (!roleDao.exists(name)) {
             throw new ValidationErrorsException("Role not found: " + name);
         }
@@ -65,10 +60,10 @@ public class RoleResourceImpl implements RoleResource, Resource {
         return new DeleteRoleResponse();
     }
 
-    private static void assertPermissions(String permission, String message) {
-        Subject subject = SecurityUtils.getSubject();
-        if (!subject.isPermitted(permission)) {
-            throw new UnauthorizedException(message);
+    private static void assertAdmin() {
+        UserPrincipal p = UserPrincipal.getCurrent();
+        if (!p.isAdmin()) {
+            throw new WebApplicationException("Only admins can do that", Response.Status.FORBIDDEN);
         }
     }
 }