+++ server/impl/src/main/java/com/walmartlabs/concord/server/user/UserResourceImpl.java
@@ -1,12 +1,9 @@
 package com.walmartlabs.concord.server.user;
 
 import com.walmartlabs.concord.server.api.PerformedActionType;
-import com.walmartlabs.concord.server.api.security.Permissions;
 import com.walmartlabs.concord.server.api.user.*;
-import org.apache.shiro.SecurityUtils;
+import com.walmartlabs.concord.server.security.UserPrincipal;
 import org.apache.shiro.authz.UnauthorizedException;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.apache.shiro.subject.Subject;
 import org.sonatype.siesta.Resource;
 import org.sonatype.siesta.Validate;
 import org.sonatype.siesta.ValidationErrorsException;
@@ -32,16 +29,15 @@ public class UserResourceImpl implements UserResource, Resource {
     @Override
     @Validate
     public CreateUserResponse createOrUpdate(CreateUserRequest request) {
+        assertAdmin();
+
         String username = request.getUsername();
 
         UUID id = userDao.getId(username);
         if (id == null) {
-            assertPermissions(Permissions.USER_CREATE_NEW, "The current user does not have permissions to create a new user");
             UserEntry e = userManager.create(username, request.getPermissions());
             return new CreateUserResponse(e.getId(), PerformedActionType.CREATED);
         } else {
-            // TODO check per-entry permissions?
-            assertPermissions(Permissions.USER_UPDATE_ANY, "The current user does not have permissions to update an existing user");
             userDao.update(id, request.getPermissions());
             return new CreateUserResponse(id, PerformedActionType.UPDATED);
         }
@@ -50,6 +46,8 @@ public class UserResourceImpl implements UserResource, Resource {
     @Override
     @Validate
     public UserEntry findByUsername(String username) {
+        assertAdmin();
+
         UUID id = userDao.getId(username);
         if (id == null) {
             throw new WebApplicationException("User not found: " + username, Status.NOT_FOUND);
@@ -58,8 +56,9 @@ public class UserResourceImpl implements UserResource, Resource {
     }
 
     @Override
-    @RequiresPermissions(Permissions.USER_DELETE_ANY)
     public DeleteUserResponse delete(UUID id) {
+        assertAdmin();
+
         if (!userDao.existsById(id)) {
             throw new ValidationErrorsException("User not found: " + id);
         }
@@ -68,10 +67,10 @@ public class UserResourceImpl implements UserResource, Resource {
         return new DeleteUserResponse();
     }
 
-    private void assertPermissions(String wildcard, String perm) {
-        Subject subject = SecurityUtils.getSubject();
-        if (!subject.isPermitted(wildcard)) {
-            throw new UnauthorizedException(perm);
+    private static void assertAdmin() {
+        UserPrincipal p = UserPrincipal.getCurrent();
+        if (!p.isAdmin()) {
+            throw new UnauthorizedException("Only admins can do that");
         }
     }
 }